Storms implemented You can configure in wolfpack.xml: - "Enable Storms" (Enable / Disable, since it make every 5 seconds two socket loops in core) - "Intensity Value to Rain Becomes Storm" (After this value for intensity, Rain is treated as a Storm. Default to 100 - Intensity goes from 16 to 112) - "Default Thunder Chance" (Its the chance for a Thunder Sound in this Region when Storm is there) - "Ray Chance on Thunder" (Its the chance for a Ray in this Region when Storm there).
About Ray:
Ray just make things brigther on screen (Like LightLevel 0) for 0,5s. Its just checked when a Thunder comes. So, make this value Higher than Thunder Chance will make every thunder comes with Rain on screen)

Last adition is about Region LightModifier. Its a parameter for LightLevel modifier for that Region. If a LightLevel in some region is 20 and LightModifier is -5, so LightLevel becomes 15 (It checks Max and Min for LightLevel after modifications).
It will be usefull for future things like make day darker in a Region where a Storm Occurs.